# Make Signals immutable, restructure annotations, clean up read and write methods