Okay, this ended up being a bigger PR than I expected -- would love to hear thoughts on any suggested changes or things I'd need to walk back here.

## Extreme makeover: EDF Edition
### EDF.jl
- Signals are now immutable
- AnnotationsList and RecordAnnotation have been restructured into a `Pair{RecordAnnotation, TimestampAnnotation}` to better reflect that the first annotation in each data record is both required and structured differently
- RecordAnnotation and TimestampAnnotation both inherit from AbstractAnnotation, allowing for dispatch during `EDF.write`
- `Header` renamed to `FileHeader` as multiple header records exist in an EDF file
- Since multiple signals contain headers, each is given a representation, allowing for clearer write and read functions

### I/O
- Functions are more finely divided into discrete sections
- Variables are given more descriptive names
- writing functions now dispatch on new annotation types

### Tests
- Pretty much updated to work with the new structs somewhat.

There's a lot here, so I decided this would probably be a good stopping point before I added anything else (and tbh I probably should have requested input a bit sooner), but I'm hoping this PR will make transitioning to an Mmap solution much easier, which, in turn, could make future EDF endeavors a bit more friendly on the memory.

### Things that still need to be added, should I get the go-ahead
- Docs
- NEW tests