@OTDE 2020-04-29
OOF. Remembering how much the me of two months ago needed to reread YASGuide. These'll be up in a sec.

> Do you know whether this has an impact on performance, e.g. @benchmark EDF.read("file.edf")

I don't, unfortunately. I could set something up in the next tagged release so that we have a way to check this, if that works!