@ararslan 2020-04-29
Should be super quick and simple to do. Check out master, run
```julia
using BenchmarkTools
using EDF
@benchmark EDF.read("some_edf_file.edf")
```
Then check out this branch and run the same thing and report both results. You could use one or more of the EDF files checked in as test cases here to benchmark.