@OTDE 2020-04-29
Old:
```
julia> @benchmark EDF.read("test/data/test.edf")
BenchmarkTools.Trial: 
  memory estimate:  3.33 MiB
  allocs estimate:  8071
  --------------
  minimum time:     937.574 μs (0.00% GC)
  median time:      962.879 μs (0.00% GC)
  mean time:        1.068 ms (7.88% GC)
  maximum time:     5.179 ms (81.16% GC)
  --------------
  samples:          4677
  evals/sample:     1
```

New:
```
@benchmark EDF.read("test/data/test.edf")
BenchmarkTools.Trial: 
  memory estimate:  3.60 MiB
  allocs estimate:  15404
  --------------
  minimum time:     1.347 ms (0.00% GC)
  median time:      1.413 ms (0.00% GC)
  mean time:        1.618 ms (9.79% GC)
  maximum time:     5.273 ms (65.65% GC)
  --------------
  samples:          3087
  evals/sample:     1
```

Looks like performance is being affected.