@OTDE 2020-04-29
After latest test:
```
julia> @benchmark EDF.read("test/data/test.edf")
BenchmarkTools.Trial: 
  memory estimate:  3.33 MiB
  allocs estimate:  6819
  --------------
  minimum time:     884.232 μs (0.00% GC)
  median time:      913.414 μs (0.00% GC)
  mean time:        1.014 ms (8.20% GC)
  maximum time:     2.654 ms (54.53% GC)
  --------------
  samples:          4924
  evals/sample:     1
```